+++ Runtime/Scripts/Messenger/MessageHolder.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using PJ.Native.Proto;
 using UnityEngine;
 
-namespace PJ.Native.Messenger
+namespace PJ.Native.PubSub
 {
     public interface MessageHolder
     {