+++ Runtime/Scripts/Messenger/MessageManager.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using PJ.Core.Util;
 using UnityEngine;
 
-namespace PJ.Native.Messenger
+namespace PJ.Native.PubSub
 {
     public class MessageManager : Singleton<MessageManager>
     {